+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Unreleased
 
 - Added output with `--verbose` option.
+- Avoided unnecessary requests to get access token.
 
 ## 0.6.4